pyproject.toml
[tool.poetry]
name = "lip2sp-pytorch"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "3.9.3"
aiosignal = "1.3.1"
antlr4-python3-runtime = "4.9.3"
appdirs = "1.4.4"
attrs = "23.2.0"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
cmake = "3.28.3"
cython = "3.0.9"
decorator = "5.1.1"
docker-pycreds = "0.4.0"
filelock = "3.13.1"
frozenlist = "1.4.1"
fsspec = "2024.2.0"
gitdb = "4.0.11"
gitpython = "3.1.42"
hydra-core = "1.3.2"
idna = "3.6"
jinja2 = "3.1.3"
lightning = {extras = ["pytorch-extra"], version = "^2.2.1"}
lightning-utilities = "0.10.1"
lit = "17.0.6"
markupsafe = "2.1.5"
mpmath = "1.3.0"
multidict = "6.0.5"
networkx = "3.2.1"
numpy = "1.26.4"
nvidia-cublas-cu11 = "11.10.3.66"
nvidia-cuda-cupti-cu11 = "11.7.101"
nvidia-cuda-nvrtc-cu11 = "11.7.99"
nvidia-cuda-runtime-cu11 = "11.7.99"
nvidia-cudnn-cu11 = "8.5.0.96"
nvidia-cufft-cu11 = "10.9.0.58"
nvidia-curand-cu11 = "10.2.10.91"
nvidia-cusolver-cu11 = "11.4.0.1"
nvidia-cusparse-cu11 = "11.7.4.91"
nvidia-nccl-cu11 = "2.14.3"
nvidia-nvtx-cu11 = "11.7.91"
omegaconf = "2.3.0"
packaging = "24.0"
pillow = "10.2.0"
protobuf = "4.25.3"
psutil = "5.9.8"
pyopenjtalk = "0.3.3"
pyreaper = "0.0.10"
pysptk = "0.2.2"
python-dotenv = "1.0.1"
pytorch-lightning = "2.2.1"
pyworld = "0.3.4"
pyyaml = "6.0.1"
requests = "2.31.0"
scipy = "1.12.0"
sentry-sdk = "1.41.0"
setproctitle = "1.3.3"
six = "1.16.0"
smmap = "5.0.1"
sympy = "1.12"
torch = "2.0.1"
torchaudio = "2.0.2"
torchmetrics = "1.3.1"
torchvision = "0.15.2"
tqdm = "4.66.2"
triton = "2.0.0"
typing-extensions = "4.10.0"
urllib3 = "2.2.1"
wandb = "0.16.4"
yarl = "1.9.4"
hydra-colorlog = "^1.2.0"
ipykernel = "^6.29.3"
timm = "^0.9.16"
cosine-annealing-warmup = {git = "https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
